Fix compilation after hercules update.

+++ src/emap/struct/itemdext.h
@@ -6,7 +6,7 @@
 
 struct ItemCardExt
 {
-    unsigned short id;
+    int id;
     unsigned short amount;
 };
